chore(vm): Add schema/types
