+++ internal/provider/vm/vm_datasource.go
@@ -26,18 +26,6 @@ func NewVMDataSource() datasource.DataSource {
 	return &vmDataSource{}
 }
 
-type VMDataSourceModel struct { //nolint:revive
-	ID          types.String `tfsdk:"id"`
-	VDC         types.String `tfsdk:"vdc"`
-	Name        types.String `tfsdk:"name"`
-	VappName    types.String `tfsdk:"vapp_name"`
-	VappID      types.String `tfsdk:"vapp_id"`
-	Description types.String `tfsdk:"description"`
-	State       types.Object `tfsdk:"state"`
-	Resource    types.Object `tfsdk:"resource"`
-	Settings    types.Object `tfsdk:"settings"`
-}
-
 type vmDataSource struct {
 	client   *client.CloudAvenue
 	vdc      vdc.VDC
